Release 1.1.1.0
Cleaned up code base.

Extracted ExecUI method into base ViewModel class to avoid having to type out that abomination of a method call in order to execute something on the UI thread...

Updated help view.

Fixed issue where joining a convo successfully via dialog (with the round button in the main UI) would not add the convo to the list.